整理角色管理

+++ src/main/java/cn/stylefeng/guns/modular/system/service/RoleService.java
@@ -1,11 +1,22 @@
 package cn.stylefeng.guns.modular.system.service;
 
+import cn.hutool.core.bean.BeanUtil;
 import cn.hutool.core.convert.Convert;
+import cn.stylefeng.guns.core.common.constant.Const;
+import cn.stylefeng.guns.core.common.constant.cache.Cache;
+import cn.stylefeng.guns.core.common.constant.factory.ConstantFactory;
+import cn.stylefeng.guns.core.common.exception.BizExceptionEnum;
 import cn.stylefeng.guns.core.common.node.ZTreeNode;
+import cn.stylefeng.guns.core.log.LogObjectHolder;
+import cn.stylefeng.guns.core.util.CacheUtil;
 import cn.stylefeng.guns.modular.system.entity.Relation;
 import cn.stylefeng.guns.modular.system.entity.Role;
 import cn.stylefeng.guns.modular.system.mapper.RelationMapper;
 import cn.stylefeng.guns.modular.system.mapper.RoleMapper;
+import cn.stylefeng.guns.modular.system.model.RoleDto;
+import cn.stylefeng.roses.core.util.ToolUtil;
+import cn.stylefeng.roses.kernel.model.exception.RequestEmptyException;
+import cn.stylefeng.roses.kernel.model.exception.ServiceException;
 import com.baomidou.mybatisplus.service.impl.ServiceImpl;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -32,13 +43,52 @@ public class RoleService extends ServiceImpl<RoleMapper, Role> {
     private RelationMapper relationMapper;
 
     /**
+     * 添加角色
+     *
+     * @author fengshuonan
+     * @Date 2018/12/23 6:40 PM
+     */
+    @Transactional(rollbackFor = Exception.class)
+    public void addRole(Role role) {
+
+        if (ToolUtil.isOneEmpty(role, role.getName(), role.getPid(), role.getDescription())) {
+            throw new RequestEmptyException();
+        }
+
+        role.setRoleId(null);
+
+        this.insert(role);
+    }
+
+    /**
+     * 编辑角色
+     *
+     * @author fengshuonan
+     * @Date 2018/12/23 6:40 PM
+     */
+    @Transactional(rollbackFor = Exception.class)
+    public void editRole(RoleDto roleDto) {
+
+        if (ToolUtil.isOneEmpty(roleDto, roleDto.getName(), roleDto.getPid(), roleDto.getDescription())) {
+            throw new RequestEmptyException();
+        }
+
+        Role old = this.selectById(roleDto.getRoleId());
+        BeanUtil.copyProperties(roleDto, old);
+        this.updateById(old);
+
+        //删除缓存
+        CacheUtil.removeAll(Cache.CONSTANT);
+    }
+
+    /**
      * 设置某个角色的权限
      *
      * @param roleId 角色id
      * @param ids    权限的id
      * @date 2017年2月13日 下午8:26:53
      */
-    @Transactional
+    @Transactional(rollbackFor = Exception.class)
     public void setAuthority(Long roleId, String ids) {
 
         // 删除该角色所有的权限
@@ -59,13 +109,29 @@ public class RoleService extends ServiceImpl<RoleMapper, Role> {
      * @author stylefeng
      * @Date 2017/5/5 22:24
      */
-    @Transactional
+    @Transactional(rollbackFor = Exception.class)
     public void delRoleById(Long roleId) {
+
+        if (ToolUtil.isEmpty(roleId)) {
+            throw new ServiceException(BizExceptionEnum.REQUEST_NULL);
+        }
+
+        //不能删除超级管理员角色
+        if (roleId.equals(Const.ADMIN_ROLE_ID)) {
+            throw new ServiceException(BizExceptionEnum.CANT_DELETE_ADMIN);
+        }
+
+        //缓存被删除的角色名称
+        LogObjectHolder.me().set(ConstantFactory.me().getSingleRoleName(roleId));
+
         //删除角色
         this.roleMapper.deleteById(roleId);
 
-        // 删除该角色所有的权限
+        //删除该角色所有的权限
         this.roleMapper.deleteRolesById(roleId);
+
+        //删除缓存
+        CacheUtil.removeAll(Cache.CONSTANT);
     }
 
     /**
